Fix file layout
