+++ src/shared/HOC/withNavbar.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import Navbar from '../../Nav/Navbar';
-import Footer from '../Footer';
+import Footer from '../Elements/Footer';
 
 const withNavbar = <P extends {}>(Component: React.ComponentType<P>) =>
   class extends React.Component<P> {
@@ -12,7 +12,7 @@ const withNavbar = <P extends {}>(Component: React.ComponentType<P>) =>
       return [
         <Navbar key={0} />,
         <Component key={1} {...this.props} />,
-        <Footer key={2} borderThickness={'1px'} />,
+        <Footer key={2} />,
       ];
     }
   };
